Update auxlib runtime.Statics to match 2.13 changes
